Conflicts:
	Collectiondb.cpp
	Collectiondb.h
	Conf.cpp
	Conf.h
	Msg39.cpp
	PageEvents.cpp
	PageResults.cpp
	PageTurk.cpp
	Pages.cpp
	Parms.cpp
	Posdb.cpp
	Proxy.cpp
	Query.cpp
	Query.h
	RdbBase.cpp
	RdbMap.cpp
	Repair.cpp
	Repair.h
	SafeBuf.cpp
	Spider.cpp
	Tagdb.cpp
	TopTree.cpp
	XmlDoc.cpp
	main.cpp

+++ Query.h
@@ -356,9 +356,11 @@ class QueryWord {
 	// . used by QueryBoolean since it uses QueryWords heavily
 	class QueryTerm *m_queryWordTerm;
 	// user defined weights
+
 	int32_t m_userWeight;
 	char m_userType;
-	int32_t m_userWeightPhrase;
+	float m_userWeightPhrase;
+
 	char m_userTypePhrase;
 	bool m_queryOp;
 	// is it after a NOT operator? i.e. NOT ( x UOR y UOR ... )
@@ -469,7 +471,7 @@ class QueryTerm {
 	// is it a repeat?
 	char m_repeat;
 	// user defined weight for this term, be it phrase or word
-	int32_t m_userWeight;
+	float m_userWeight;
 	char m_userType;
 	// . is this query term before a | (pipe) operator?
 	// . if so we must read the whole termlist, like m_underNOT above
